Reorganized. Implemented icon loader

+++ src/com/alteredmechanism/java/awt/FontFactory.java
@@ -7,16 +7,16 @@ import java.awt.Toolkit;
 public class FontFactory {
 
 	public Font create(String name, int style, int size) {
-		return new Font(name, style, pointToPixel(size));
+		return new Font(name, style, (int) pointToPixel(size));
 	}
 	
 	public Font derive(Font f) {
 		return f.deriveFont(pointToPixel(f.getSize()));
 	}
 
-    public int pointToPixel(float pointSize) {
+    public float pointToPixel(float pointSize) {
         float ppi = Toolkit.getDefaultToolkit().getScreenResolution();
-        int pixelHeight = (int) Math.round(pointSize / (72f / ppi));
+        float pixelHeight = Math.round(pointSize / (72f / ppi));
         System.out.println("Converted font point size " + pointSize + " to pixel height = " + pixelHeight);
         return pixelHeight;
     }
